libmariadb-dev-compat must conflict libmariadbclient-dev/-compat
The package libmariadbclient-dev-compat provide most of the same files
(e.g. libmysqlclient.so) and the libmariadbclient-dev also provides some
files (e.g. mysql_config), so they must conflict each other.

This set of packages provide the same functionality, so it is logical
that they conflict and cannot be installed at the same time.